net-test: packetdrill: Extend data offset.
If the EDO Extension option is already used and the length
of the total options overflows 40 bytes, we re-allocate
memory for TCP options.

We save the header length in struct packet and use it in
packet_tcp_header_len() to print the user data length
properly.

+++ gtests/net/packetdrill/tcp_options.h
@@ -73,8 +73,8 @@ struct edo {
 /* Represents a list of TCP options in their wire format. */
 struct tcp_options {
 	u8 *data;		/* The options data, in wire format */
-	u8 max;			/* Max length, in bytes, of the data */
-	u8 length;		/* The length, in bytes, of the data */
+	u16 max;		/* Max length, in bytes, of the data */
+	u16 length;		/* The length, in bytes, of the data */
 
 	struct edo *edo;
 	bool auto_hdr;
